feature: token and user extractor

+++ blog-backend/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router();
-const Blog = require('../models/models.js');
-const User = require('../models/user.js');
+const Blog = require('../models/models');
+const User = require('../models/user');
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -10,7 +10,12 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
 
-    const user = await User.findById(body.userId);
+    const decodedToken = request.user;
+    if (!decodedToken || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
+    const user = await User.findById(decodedToken.id);
     if (!user) {
         return response.status(400).json({ error: 'user not found' });
     }
@@ -28,7 +33,7 @@ blogsRouter.post('/', async (request, response) => {
     });
     const savedBlog = await blog.save()
 
-    user.blogs = user.blogs.concat(blog._id);
+    user.blogs = user.blogs.concat(savedBlog._id);
     await user.save();
 
     response.status(201).json(savedBlog);
@@ -41,12 +46,22 @@ blogsRouter.get('/:id', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' });
+    }
+    if (blog.user.toString() !== request.user.id) {
+        return response.status(401).json({ error: 'user not authorized' });
+    }
     await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
     const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' });
+    }
     blog.likes = request.body.likes;
     await blog.save();
     response.json(blog);
